Added_Destroy_Method
I had the need to destroy this plugin properly. This came from the
following scenario. I had a Light Box that occupied the full screen,
but launched it from the bottom of a page, this plugin would have
produced a button with a z-index high enough to put it ontop of my
light box. So, I need to make it go away properly.